% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDiagramGeneric.R
\name{drawDiagramGeneric}
\alias{drawDiagramGeneric}
\title{drawDiagramGeneric}
\usage{
drawDiagramGeneric(n, go, nogo, xmax = NULL, ymax = NULL, ylab = NULL)
}
\description{
drawDiagramGeneric
}
\examples{
 go <- cbind(6:8, rep(8,3))
 nogo <- cbind(0:5, rep(8,6))
 drawDiagramGeneric(n=8, go=go, nogo=nogo, ylab="Number of successes")
}
\author{
Martin Law, \email{martin.law@mrc-bsu.cam.ac.uk}
}
