% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find2stageThetaF.R
\name{find2stageThetaF}
\alias{find2stageThetaF}
\title{Find conditional power for futility}
\usage{
find2stageThetaF(r, r1, n2, p)
}
\arguments{
\item{r}{Number of responses to be exceeded for trial success}

\item{r1}{Number of responses to be exceeded to continue to second stage}

\item{n2}{Sample size of second stage}

\item{p}{Response rate under alternative hypothesis H1}
}
\value{
The maximum conditional power at which the trial will stop for futility. This is the CP when there are r1 responses out of n1 participants.
}
\description{
Finds maximum conditional power at which trial will stop for futility
}
\examples{
find2stageThetaF(r=20, r1=8, n2=13, p=0.92)

}
